+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+- [#21] Added support for `--verbose`/`-v` to `get:status`
+
 ### 0.6.0
 
 - [#17] Added `get:release-line` command
@@ -31,3 +33,4 @@
 [#14]: https://github.com/warehouseai/wrhs/pull/14
 [#16]: https://github.com/warehouseai/wrhs/pull/16
 [#17]: https://github.com/warehouseai/wrhs/pull/17
+[#21]: https://github.com/warehouseai/wrhs/pull/21